Moved provider_info_endpoint from extensions to oauth
Ref #638
