+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on the [KeepAChangeLog] project.
 ### Added
 - [#655] Host can be forced on webfinger discovery
 - [#441] CookieDealer now accepts secure and httponly params
+- [#638] Moved `providerinfo_endpoint` from `oic.extensions` to `oic.oauth2`
 
 [#655]: https://github.com/OpenIDC/pyoidc/issues/655
 [#598]: https://github.com/OpenIDC/pyoidc/issues/598
@@ -52,6 +53,7 @@ The format is based on the [KeepAChangeLog] project.
 [#642]: https://github.com/OpenIDC/pyoidc/pull/642
 [#639]: https://github.com/OpenIDC/pyoidc/issues/639
 [#631]: https://github.com/OpenIDC/pyoidc/issues/631
+[#638]: https://github.com/OpenIDC/pyoidc/issues/638
 
 ## 0.15.1 [2019-01-31]
